Refactor client side components to new component to allow for async server calls to CMS for posters.
